Bugfix | Adjust file upload communication between front and processing
* atoms folder structure

* add validation for modal and remove file from uploads on component unmount

* empty file big fix

* bottom sidebar links fix

* add few or more names to mapping

* correct data send to processing and correct state in ui

* remove files inside /uploads from .gitignore

* add jsdoc to csv parse functionality

* change regex pattern

* add check if mimetype is ok and parse xls,xlsx

* add validation for mimetype

* install xlsx package

* handle success and error on file delete

+++ qk-front/pages/api/file-delete.js
@@ -3,7 +3,11 @@ import fs from "fs"
 export default async (req, res) => {
    if (req.method === "POST") {
       const filePath = `uploads/${req.body}`
-      fs.unlinkSync(filePath)
-      res.status(200)
+      if (fs.existsSync(filePath)) {
+         fs.unlinkSync(filePath)
+         res.status(200).send("OK")
+      } else {
+         res.status(500).statusText("Something went wrong")
+      }
    }
 }
